requirements.txt
selenium
beautifulsoup4
schedule
facebook-sdk
webdriver-manager
